ATM-1302: Implement Integration Tests for POST /rest/api/2/issue Endpoint
