+++ src/api/routes/issueRoutes.create.persistence.success.test.ts
@@ -410,167 +410,4 @@ describe('POST /rest/api/2/issues - Persistence Verification', () => { // Update
     expect(persistedParent.childIssueKeys.length).toBe(1); // Should only contain the newly created subtask
   });
 
-  // --- New Tests for Invalid Subtask Parent ---
-
-  // Test Case 2: Create Subtask with Non-Existent Parent Key - Persistence Verification
-  it('should return 404 and not persist the Subtask issue when parent key does not exist', async () => {
-      // Arrange
-      const nonExistentParentKey = 'NON-EXISTENT-123';
-      const reqBody = {
-          fields: {
-              issuetype: { name: 'Subtask' },
-              summary: 'Subtask with non-existent parent',
-              parent: { key: nonExistentParentKey },
-          }
-      };
-
-      // Act
-      const response = await request(app).post('/rest/api/2/issues').send(reqBody);
-
-      // Assert (API response)
-      expect(response.status).toBe(404); // Expecting 404 if parent key is not found
-      expect(response.body).toHaveProperty('error'); // Expecting an error message in the body
-      expect(response.body.error).toContain(`Parent issue with key '${nonExistentParentKey}' not found`); // More specific error message check
-
-      // Assert (Persistence)
-      const dbContent = await fs.promises.readFile(dbPath, 'utf8');
-      const dbData = JSON.parse(dbContent);
-
-      // Check that no issue was added to the database
-      expect(dbData.issues.length).toBe(0);
-      // Check that the issue key counter hasn't increased (it should still be 0 because beforeEach sets it)
-      expect(dbData.issueKeyCounter).toBe(0);
-  });
-
-  // Test Case 3: Create Subtask with Parent of Invalid Type (e.g., Bug) - Persistence Verification
-  it('should return 400 and not persist the Subtask issue when parent is a Bug', async () => {
-      // Arrange: Create a Bug issue first
-      const bugReqBody = {
-          fields: {
-              issuetype: { name: 'Bug' },
-              summary: 'Invalid parent Bug',
-          }
-      };
-      const bugResponse = await request(app).post('/rest/api/2/issues').send(bugReqBody);
-      expect(bugResponse.status).toBe(201);
-      const bugKey = bugResponse.body.key;
-      expect(bugKey).toBeDefined();
-
-      // Construct Subtask request body pointing to the Bug parent
-      const subtaskReqBody = {
-          fields: {
-              issuetype: { name: 'Subtask' },
-              summary: 'Subtask with Bug parent',
-              parent: { key: bugKey },
-          }
-      };
-
-      // Act
-      const response = await request(app).post('/rest/api/2/issues').send(subtaskReqBody);
-
-      // Assert (API response)
-      expect(response.status).toBe(400); // Expecting 400 for invalid parent type
-      expect(response.body).toHaveProperty('error'); // Expecting an error message in the body
-      expect(response.body.error).toContain(`Issue with key '${bugKey}' has type 'Bug', which cannot be a parent of a Subtask`); // Specific error message check
-
-      // Assert (Persistence)
-      const dbContent = await fs.promises.readFile(dbPath, 'utf8');
-      const dbData = JSON.parse(dbContent);
-
-      // Check that only the original Bug issue exists
-      expect(dbData.issues.length).toBe(1);
-      const persistedBug = dbData.issues.find((issue: AnyIssue) => issue.key === bugKey);
-      expect(persistedBug).toBeDefined(); // The Bug should still be there
-      expect(persistedBug.issueType).toBe('Bug'); // Check its type
-      // Check that the Bug issue's childIssueKeys was NOT updated (it's not an Epic, so it likely remains empty or undefined)
-      // The createIssueService sets childIssueKeys to [] for non-Epics initially.
-      expect(persistedBug).toHaveProperty('childIssueKeys');
-      expect(Array.isArray(persistedBug.childIssueKeys)).toBe(true);
-      expect(persistedBug.childIssueKeys.length).toBe(0); // Should still be empty
-
-      // Check that the Subtask was NOT added
-      const persistedSubtask = dbData.issues.find((issue: AnyIssue) => issue.summary === 'Subtask with Bug parent');
-      expect(persistedSubtask).toBeUndefined(); // The subtask should NOT be found
-
-      // Check that the issue key counter reflects only the created Bug issue
-      expect(dbData.issueKeyCounter).toBe(1); // Assumes the counter started at 0 and incremented once for the Bug
-  });
-
-  // Test Case 4: Create Subtask with Parent of Invalid Type (e.g., Subtask) - Persistence Verification
-   it('should return 400 and not persist the second Subtask issue when parent is another Subtask', async () => {
-      // Arrange: Create a valid parent (e.g., Epic) and a first Subtask linked to it
-      const epicReqBody = {
-          fields: {
-              issuetype: { name: 'Epic' },
-              summary: 'Valid parent Epic',
-          }
-      };
-      const epicResponse = await request(app).post('/rest/api/2/issues').send(epicReqBody);
-      expect(epicResponse.status).toBe(201);
-      const epicKey = epicResponse.body.key;
-      expect(epicKey).toBeDefined();
-
-      const firstSubtaskReqBody = {
-          fields: {
-              issuetype: { name: 'Subtask' },
-              summary: 'First Subtask under Epic',
-              parent: { key: epicKey },
-          }
-      };
-      const firstSubtaskResponse = await request(app).post('/rest/api/2/issues').send(firstSubtaskReqBody);
-      expect(firstSubtaskResponse.status).toBe(201);
-      const firstSubtaskKey = firstSubtaskResponse.body.key;
-      expect(firstSubtaskKey).toBeDefined();
-
-      // Verify initial state: Epic and First Subtask exist, Epic has First Subtask as child
-      let dbContentInitial = await fs.promises.readFile(dbPath, 'utf8');
-      let dbDataInitial = JSON.parse(dbContentInitial);
-      expect(dbDataInitial.issues.length).toBe(2);
-      const persistedEpicInitial = dbDataInitial.issues.find((issue: AnyIssue) => issue.key === epicKey);
-      expect(persistedEpicInitial).toBeDefined();
-      expect(persistedEpicInitial.childIssueKeys).toContain(firstSubtaskKey);
-      expect(dbDataInitial.issueKeyCounter).toBe(2); // Counter incremented twice
-
-      // Construct the second Subtask request body, pointing to the first Subtask as parent
-      const secondSubtaskReqBody = {
-          fields: {
-              issuetype: { name: 'Subtask' },
-              summary: 'Second Subtask under First Subtask (Invalid Parent)',
-              parent: { key: firstSubtaskKey }, // Invalid parent key (a Subtask)
-          }
-      };
-
-      // Act
-      const response = await request(app).post('/rest/api/2/issues').send(secondSubtaskReqBody);
-
-      // Assert (API response)
-      expect(response.status).toBe(400); // Expecting 400 for invalid parent type
-      expect(response.body).toHaveProperty('error'); // Expecting an error message in the body
-      expect(response.body.error).toContain(`Issue with key '${firstSubtaskKey}' has type 'Subtask', which cannot be a parent of a Subtask`); // Specific error message check
-
-      // Assert (Persistence)
-      const dbContent = await fs.promises.readFile(dbPath, 'utf8');
-      const dbData = JSON.parse(dbContent);
-
-      // Check that the database state is unchanged from the initial state (before the failed creation attempt)
-      expect(dbData.issues.length).toBe(2); // Still only the Epic and the first Subtask
-      const persistedEpic = dbData.issues.find((issue: AnyIssue) => issue.key === epicKey);
-      expect(persistedEpic).toBeDefined();
-      expect(persistedEpic.childIssueKeys).toContain(firstSubtaskKey);
-      expect(persistedEpic.childIssueKeys.length).toBe(1); // Epic still only has the first Subtask as child
-
-      const persistedFirstSubtask = dbData.issues.find((issue: AnyIssue) => issue.key === firstSubtaskKey);
-      expect(persistedFirstSubtask).toBeDefined();
-      expect(persistedFirstSubtask.issueType).toBe('Subtask');
-      expect(persistedFirstSubtask.parentKey).toBe(epicKey); // First Subtask still points to Epic
-
-      // Check that the second Subtask was NOT added
-      const persistedSecondSubtask = dbData.issues.find((issue: AnyIssue) => issue.summary === 'Second Subtask under First Subtask (Invalid Parent)');
-      expect(persistedSecondSubtask).toBeUndefined(); // The second subtask should NOT be found
-
-      // Check that the issue key counter has not increased further
-      expect(dbData.issueKeyCounter).toBe(2); // Counter should still be 2
-  });
-
-
 }); // End of describe Persistence Verification
